comment fix

+++ src/main/java/com/clickpost/shippingplatform/service/ordercreation/object/ShipmentItemInfo.java
@@ -7,12 +7,12 @@ public class ShipmentItemInfo {
     private final String sku_;
     private final String description_;
     private final int quantity_;
-    private final int price_;
+    private final float price_;
 
     @Nullable
     private final ShipmentItemAdditionalInfo shipmentItemAdditionalInfo_;
 
-    public ShipmentItemInfo(String sku, String description, int quantity, int price,
+    public ShipmentItemInfo(String sku, String description, int quantity, float price,
                             ShipmentItemAdditionalInfo shipmentItemAdditionalInfo) {
         sku_ = sku;
         description_ = description;
@@ -33,7 +33,7 @@ public class ShipmentItemInfo {
         return quantity_;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price_;
     }
